+++ src/components/Selector/Selector.tsx
