+++ src/domains/ShoppingProducts/components/CartProduct/CartProduct.style.ts
