+++ src/domains/ShoppingProducts/components/Modal/Modal.tsx
@@ -1,4 +1,4 @@
-import { useKeyPress } from "../../../hook/useKeyPress";
+import { useKeyPress } from "../../../../hook/useKeyPress";
 import {
   ModalBackdrop,
   ModalContainer,
