+++ src/domains/ShoppingProducts/components/Product/Product.style.ts
