+++ src/domains/ShoppingProducts/components/Product/Product.tsx
