+++ src/domains/ShoppingProducts/components/TitleContainer/TitleContainer.style.ts
